Move internal/pkg/provider to pkg/provider
