+++ cmd/cofidectl/cmd/down.go
@@ -10,7 +10,7 @@ import (
 	trust_zone_proto "github.com/cofide/cofide-api-sdk/gen/go/proto/trust_zone/v1alpha1"
 	cmdcontext "github.com/cofide/cofidectl/cmd/cofidectl/cmd/context"
 	"github.com/cofide/cofidectl/cmd/cofidectl/cmd/statusspinner"
-	"github.com/cofide/cofidectl/internal/pkg/provider/helm"
+	"github.com/cofide/cofidectl/pkg/provider/helm"
 	"github.com/spf13/cobra"
 )
